backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
httpx==0.27.2
beautifulsoup4==4.12.3
openpyxl==3.1.5
python-dotenv==1.0.1
rapidfuzz==3.9.7
websockets==13.1
python-multipart==0.0.9
